Moved getInitialQuery from FindUtils to FindBar
